[MOD] Refactor A1 package i cadenes
