+++ src/main/java/ed/sergi/Persona.java
@@ -1,4 +1,4 @@
-package ed.arturo;
+package ed.sergi;
 
 public class Persona {
